created linked and sensitivity versions of model (work in progress) and versions of model with alternative business models.
The alternative business models change the benefits of the NPV formulations. Demand Charge has the benefit of the electricity that can be directly used from the panels (instead of all of the electricity that is generated by the panels, which is the benefit under a net-metering system). WholesaleComp uses the wholesale price of electricity to compensate PV customers.

The import sheet has also been extensively updated, especially the sensitivity variables. Still need to whittle these down to make a reasonable number of evaluation cases.